+++ test/testCommon.ts
@@ -5,6 +5,7 @@ export const configuration = process.env.configuration || 'debug';
 export const stagingDirectory = path.join(__dirname, '..', '..', 'lib', configuration);
 
 export enum TestConstants {
+    Error = 'Error',
     Success = 'Success',
     TaskTestTrace = 'TASK_TEST_TRACE',
     MockReturnCode = 'MOCK_RETURN_CODE',
